requirements.txt
2captcha-python
loguru
Pillow
numpy